\name{accum}
\alias{accum}
\title{Response calculation}
\usage{
  accum(dwide)
}
\arguments{
  \item{dwide}{iTRAQ data in wide format including columns
  corresponding to iTRAQ channels containing their
  intensities.}
}
\description{
  Calculates needed sample size accumulation from iTRAQ
  data which is given on spectrum level.
}

