\name{addIonSatistics}
\alias{addIonSatistics}
\title{Summary statistics -- Ion intensities per spectra}
\usage{
  addIonSatistics(dwide)
}
\arguments{
  \item{dwide}{iTRAQ data in wide format including columns
  corresponding to iTRAQ channels containing their
  intensities.}
}
\description{
  Summary statistics -- Ion intensities per spectra
}

