\name{addLoadings}
\alias{addLoadings}
\title{Adjust for confounding -- add an appropiate  target}
\usage{
  addLoadings(dwide, byRef = F)
}
\arguments{
  \item{dwide}{iTRAQ data in wide format}

  \item{byRef}{schould the average be calculated from the
  loading of the reference channel. Default is FALSE and
  this is recommended.}
}
\description{
  Adjust for confounding -- add an appropiate target
}

