\name{addRetentionIndexTimeStatistics}
\alias{addRetentionIndexTimeStatistics}
\title{Summary statistics -- Calculates index retention time statistics}
\usage{
  addRetentionIndexTimeStatistics(dwide, ...)
}
\arguments{
  \item{dwide}{iTRAQ data in wide format}

  \item{...}{Additional arguments passed for ddply}
}
\description{
  Summary statistics -- Calculates index retention time
  statistics
}

