\name{adjusting}
\alias{adjusting}
\title{Adjust for confounding -- State of the art adjustments for confounding}
\usage{
  adjusting(dwide)
}
\arguments{
  \item{dwide}{iTRAQ data in wide format.}
}
\description{
  Compensate for possible confounding due the transformed
  values for the statistical analysis.
}

