\name{channelResponses}
\alias{channelResponses}
\title{Response calculation}
\usage{
  channelResponses(dwide, acc)
}
\arguments{
  \item{dwide}{iTRAQ data in wide format including columns
  corresponding to iTRAQ channels containing their
  intensities.}

  \item{acc}{result of an accumulation of sample sizes}
}
\description{
  From spectrum to protein level -- Response variable
  calculation
}

