\name{pVolcano}
\alias{pVolcano}
\title{Volcano plot}
\usage{
  pVolcano(res, threshold, .foldchange = TRUE,
    .plot = TRUE)
}
\arguments{
  \item{res}{result frame of test results}

  \item{threshold}{for biological reasonable effect}

  \item{.foldchange}{wheather results given in ratios or
  log-ratios}

  \item{.plot}{if true adds a plotting layer}
}
\description{
  Volcano plot
}

