% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{reformat1}
\alias{reformat1}
\title{Reformatting original proteomics csv file}
\usage{
reformat1(data, k)
}
\arguments{
\item{data}{Original csv file.}

\item{k}{The number of observations.}
}
\value{
reformatted csv file.
}
\description{
Reformatting original proteomics csv file
}
\examples{
setwd("C:/Users/kbs/Desktop/RA_Pig")
# data = original data in .csv format 
data<-read.csv("LI-A-2.csv")
}
