% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reformat2.R
\name{reformat2}
\alias{reformat2}
\title{Reformatting the original proteomics csv file}
\usage{
reformat2(data)
}
\arguments{
\item{data}{reformat1 output csv file:reformat1 output contains protein and peptide in the same row.}
}
\value{
protein info and peptide info are separated.
}
\description{
Reformatting the original proteomics csv file
}
\examples{
setwd("C:/Users/kbs/Desktop/RA_pig2/Liver_samples")
data<-read.csv("Control_Liver.csv")
data2<-reformat2(data)
}
