% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Repeated_peptide.R
\name{Repeated_peptide2}
\alias{Repeated_peptide2}
\title{A set of Repeated peptides abundances.}
\usage{
Repeated_peptide2(data, first_col, last_col)
}
\arguments{
\item{data}{Reformatted csv file from the reformat1 function.}

\item{first_col}{The first column of observations.}

\item{last_col}{The last column of observations.}
}
\value{
A set of Repeated peptides abundances.
}
\description{
A set of Repeated peptides abundances.
}
\examples{
# warning: I assumed that a vector of obserations are the same if they have the same observations.??
setwd("C:/Users/kbs/Desktop/RA_pig2")
Control_Liver<-read.csv("Control_Liver.csv")
Infected_Liver<-read.csv("Infected_Liver.csv")
Repeated_peptide2(Control_Liver,3,14)

}
