\name{responseStatisics}
\alias{responseStatisics}
\title{Summary statistics -- Generic to calculate summary statistics}
\usage{
  responseStatisics(dwide, frm)
}
\arguments{
  \item{dwide}{iTRAQ data in wide format}

  \item{frm}{for example: frm <- value ~ protein + variable
  frm <- value ~ peptide + variable}
}
\description{
  Calculates generic summary statistics based on a given
  formula.
}

