\name{selectByEffect}
\alias{selectByEffect}
\title{Result filtering -- Test for biological effect}
\usage{
  selectByEffect(res, cutoff = 1)
}
\arguments{
  \item{res}{Result file}

  \item{cutoff}{the cutoff to be used in the selection}
}
\description{
  Result filtering -- Test for biological effect
}

