\name{testForPeptideEffect}
\alias{testForPeptideEffect}
\title{Data Analysis -- Testing on peptide level}
\usage{
  testForPeptideEffect(dat, frm, conf.level, ...)
}
\arguments{
  \item{dat}{iTRAQ data in long format}

  \item{frm}{formal for the test}

  \item{conf.level}{confidence level}

  \item{...}{arguments understood by ddply}
}
\description{
  Data Analysis -- Testing on peptide level
}

