\name{testForProteinEffect}
\alias{testForProteinEffect}
\title{Data Analysis -- Testing on protein level}
\usage{
  testForProteinEffect(dat, frm, conf.level, ...)
}
\arguments{
  \item{dat}{iTRAQ data in long format}

  \item{frm}{formal for the test}

  \item{conf.level}{confidence level}

  \item{...}{arguments understood by ddply}
}
\description{
  Data Analysis -- Testing on protein level
}

