\name{testing}
\alias{testing}
\title{Data Analysis -- Testing features with Tukey Honest Significant Differences}
\usage{
  testing(dp, frm, conf.level)
}
\arguments{
  \item{dp}{iTRAQ data in long format}

  \item{frm}{formal for the test}

  \item{conf.level}{confidence level}
}
\description{
  Data Analysis -- Testing features with Tukey Honest
  Significant Differences
}

