\name{testingOneshot}
\alias{testingOneshot}
\title{Data Analysis -- Testing one feature without Tukey Honest Significant Differences}
\usage{
  testingOneshot(model)
}
\arguments{
  \item{model}{ANOVA model of the corresponding fit}
}
\description{
  Data Analysis -- Testing one feature without Tukey Honest
  Significant Differences
}

