\name{testingTukey}
\alias{testingTukey}
\title{Data Analysis -- Testing one feature with Tukey Honest Significant Differences}
\usage{
  testingTukey(model, conf.level)
}
\arguments{
  \item{model}{ANOVA model of the corresponding fit}

  \item{conf.level}{confidence level}
}
\description{
  Data Analysis -- Testing one feature with Tukey Honest
  Significant Differences
}

