% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Repeated_peptide.R
\name{W_Repeated_peptide}
\alias{W_Repeated_peptide}
\title{Save all repeated peptide names and abundances after removing totally missing abundances.}
\usage{
W_Repeated_peptide(name, first_col, last_col)
}
\arguments{
\item{name}{File name.}

\item{first_col}{The first column of observations.}

\item{last_col}{The last column of observations.}
}
\value{
Write a set of Repeated peptides abundances in csv format.(_repeated.csv)
}
\description{
Save all repeated peptide names and abundances after removing totally missing abundances.
}
\examples{
setwd("C:/Users/kbs/Desktop/RA_pig2")
W_Repeated_peptide(Control_Liver)

}
