% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reformat1.R
\name{reformat1}
\alias{reformat1}
\title{Reformatting the original proteomics csv file}
\usage{
reformat1(data, k, s1, s2)
}
\arguments{
\item{data}{Original csv file.}

\item{k}{The number of observations.}

\item{s1}{The starting column number of protein abundance.}

\item{s2}{The starting column number of peptide abundance.}
}
\value{
reformatted csv file.
}
\description{
Reformatting the original proteomics csv file
}
\examples{
setwd("C:/Users/kbs/Desktop/RA_pig2/Liver_samples")
data<-read.csv("Infected-4.csv")
temp<-reformat1(data,12,50,50)
}
